Added Camera Scan Setup(Not finished)
Added functionality to open AQC from anilox, with the potential of setting the number of sections and scans. Waiting for Jon to make a software change request, as wires were crossed as to the functionality needed. Will finish once the change request has been completed and approved and make AMS MS have the same option.